CoOS 1.12 and 1.1.6 (latest), and FreeRTOS 8.2.1
Ported FreeRTOS 8.2.1 by finding the changes in the 7.0.1 version and
carrying them over. Then found an almost working port of CoOS, and
corrected it to work, and then carried the changes to 1.1.6 Tested all
of them with USBSerial, GPIO, SPI, DMA and a DMA ISR all working fine.
Similar examples ported to all of them.