upload-reset: fix/rewrite
The code in upload-reset did not work for me, and quick googling
shows I'm not the only one. I tried to fix it (and succeeded)
and in the process I rewrote it to be way simpler/cleaner.

Most probably the code was not working because TIOCMGET and then
TIOCMSET were used, which is slower than use a direct ioctl command
to set/clear a flag. Maybe there were something else.

Anyway, the new code is working almost always for my Blue Pill.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>