Adds support for 16Mhz thru menus
Only a modificatio nto the core (rcc.h) to support HSE/2 (needed when
using the 16Mhz crystal, to divide /2 before entering the PLL).
Tested to work with existing 8Mhz boards as well as boards with 16Mhz.
Adds a simple way to use a 16Mhz crystal thru menus in board.txt
The change is applied to the R variant to use with the bluepillF4 board
with F1 RxT mcus.