Remove libc exit and atexit handlers
NOTE: Original Arduino cores also replace atexit with dummy handler.

exit()/atexit() handling is *only* for when main exits, which is not
possible in the Arduino environment.

We use the gcc linker to mangle these symbols and replace them with
our own library.  Symbols are stong in the default library, so simple
overloading does not work.

Total saving around 5k of flash and 2k of ram - all for functionality
which is linked in but never used.

* all patches are independant and functional, but most saving is with
all patches included.